requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Langchain Classic
langchain-classic

# Langchain Community
langchain-community

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
jinja2

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn